+++ FashionToday_App/Pashion_Today/app/build.gradle
@@ -8,11 +8,11 @@ android {
     compileSdkVersion 29
     buildToolsVersion "29.0.0"
     defaultConfig {
-        applicationId "com.example.pashion_today"
+        applicationId "com.soma.pashion_today"
         minSdkVersion 15
         targetSdkVersion 29
-        versionCode 1
-        versionName "1.0"
+        versionCode 2
+        versionName "1.1"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
@@ -43,4 +43,5 @@ dependencies {
     androidTestImplementation 'androidx.test:runner:1.2.0'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
     implementation("com.squareup.okhttp3:okhttp:4.0.1")
+    implementation 'com.prolificinteractive:material-calendarview:1.4.3'
 }
