+++ FashionToday_App/Pashion_Today/app/src/main/res/layout/calendar_do.xml
@@ -1,0 +1,68 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+              android:orientation="vertical"
+              android:layout_width="match_parent"
+              android:layout_height="match_parent">
+    <TextView
+            android:text="일정 등록"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content" android:layout_alignParentStart="true"
+            android:layout_marginStart="154dp" android:layout_alignParentLeft="true"
+            android:layout_alignParentTop="true" android:id="@+id/textView2" android:layout_marginTop="41dp"
+            android:layout_marginLeft="154dp" android:textAllCaps="true"
+            android:textAppearance="@style/TextAppearance.AppCompat.Large" android:textStyle="bold"
+            android:layout_centerHorizontal="true"/>
+    <TextView
+            android:text="날짜 : "
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content" android:layout_alignParentStart="true"
+            android:layout_marginStart="38dp" android:layout_alignParentLeft="true" android:layout_alignParentTop="true"
+            android:id="@+id/textView3" android:layout_marginTop="147dp" android:layout_marginLeft="38dp"
+            android:textAppearance="@style/TextAppearance.AppCompat.Large"/>
+    <TextView
+            android:text="위치 : "
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content" android:layout_alignParentStart="true"
+            android:layout_marginStart="38dp" android:layout_alignParentLeft="true"
+            android:id="@+id/textView4" android:layout_marginTop="68dp" android:layout_marginLeft="38dp"
+            android:layout_below="@+id/textView3" android:textAppearance="@style/TextAppearance.AppCompat.Large"/>
+    <TextView
+            android:text="일정 : "
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content" android:layout_alignParentStart="true"
+            android:layout_marginStart="38dp" android:layout_alignParentLeft="true"
+            android:id="@+id/textView5" android:layout_marginLeft="38dp"
+            android:textAppearance="@style/TextAppearance.AppCompat.Large"
+            android:layout_marginTop="37dp"
+            android:layout_below="@+id/textView4"/>
+    <DatePicker android:layout_width="278dp"
+                android:layout_height="149dp"
+                android:datePickerMode="spinner"
+                android:calendarViewShown="false" android:layout_marginBottom="43dp"
+                android:layout_marginEnd="27dp"
+                android:visibility="visible" android:layout_alignParentEnd="true"
+                android:layout_alignParentRight="true" android:layout_marginRight="27dp"
+                android:layout_toEndOf="@+id/textView3" android:layout_toRightOf="@+id/textView3"
+                android:layout_marginLeft="12dp" android:layout_marginStart="12dp" android:layout_marginTop="17dp"
+                android:layout_below="@+id/textView2" android:id="@+id/date_picker"/>
+    <EditText
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:inputType="textPersonName"
+            android:ems="10"
+            android:layout_alignParentRight="true" android:layout_alignParentEnd="true"
+            android:layout_alignParentTop="true" android:layout_marginEnd="86dp" android:id="@+id/editText3"
+            android:layout_marginTop="238dp" android:layout_marginRight="86dp" android:hint="강남 선릉역"
+            android:layout_toEndOf="@+id/textView4" android:layout_toRightOf="@+id/textView4"
+            android:layout_marginLeft="20dp" android:layout_marginStart="18dp"/>
+    <EditText
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:inputType="textPersonName"
+            android:ems="10"
+            android:layout_alignParentRight="true" android:layout_alignParentEnd="true" android:layout_marginEnd="84dp"
+            android:id="@+id/editText5"
+            android:layout_marginRight="84dp" android:layout_marginTop="22dp" android:layout_below="@+id/editText3"
+            android:hint="데이트" android:layout_toEndOf="@+id/textView5" android:layout_toRightOf="@+id/textView5"
+            android:layout_marginLeft="20dp" android:layout_marginStart="20dp"/>
+</RelativeLayout>
