--- FashionToday_App/Pashion_Today/app/src/main/res/menu/closet_camera.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-      xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item android:title="my_page" android:orderInCategory="101" android:icon="@android:drawable/ic_menu_myplaces"
-          app:showAsAction="always" android:id="@+id/my_page_menu"/>
-    <item android:title="camera" android:id="@+id/camera_menu" android:icon="@drawable/ic_menu_camera"
-          app:showAsAction="always"/>
-    <item android:icon="@android:drawable/ic_menu_info_details" app:showAsAction="always"
-          android:id="@+id/notice_menu" android:title="notice"/>
-</menu>
